добавил модели

+++ database/migrations/2020_02_12_201438_create_agreement_table.php
@@ -19,7 +19,7 @@ class CreateAgreementTable extends Migration
             $table->text("description")->nullable();
             $table->integer("version")->default(1);
             $table->string("status"); // ["active", "new", "old"]
-            $table->integer("version_owner_user_id");
+            $table->bigInteger("version_owner_user_id");
             $table->foreign('version_owner_user_id')->references('id')->on('users');
             $table->timestamps();
         });
